fix:correct the tipe of data required

+++ database/migrations/2025_02_19_101647_create_airships_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('airships', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigint('seats');
+            $table->Integer('seats');
             $table->timestamps();
         });
     }
